+++ map_frame_session.h
@@ -30,9 +30,9 @@ struct MapFrameSessionContext {
     ImVec2 size = ImVec2(0.0f, 0.0f);
     double* center_lon = nullptr;
     double* center_lat = nullptr;
-    int* zoom_ptr = nullptr;
+    double* zoom_ptr = nullptr;
     int max_zoom = 0;
-    int zoom = 0;
+    double zoom = 0.0;
     int math_zoom = 0;
     float zoom_scale = 1.0f;
     int lod_ring_step = 1;
